+++ server/tests/src/main/scala/sttp/tapir/server/tests/ServerOptionsTests.scala
@@ -8,8 +8,11 @@ import sttp.model._
 import sttp.monad.MonadError
 import sttp.tapir._
 import sttp.tapir.server.interceptor.CustomiseInterceptors
+import sttp.tapir.server.interceptor.log.DefaultServerLog
 import sttp.tapir.server.model.ValuedEndpointOutput
 import sttp.tapir.tests._
+
+import java.util.concurrent.atomic.AtomicInteger
 
 class ServerOptionsTests[F[_], OPTIONS, ROUTE](
     createServerTest: CreateServerTest[F, Any, OPTIONS, ROUTE],
@@ -38,6 +41,69 @@ class ServerOptionsTests[F[_], OPTIONS, ROUTE](
       basicRequest.get(uri"$baseUri/incorrect").send(backend).map { response =>
         response.code shouldBe StatusCode.NotFound
         response.body shouldBe Left("ERROR: Not Found")
+      }
+    }, {
+      // https://github.com/softwaremill/tapir/issues/2827
+      val logHandledCounter = new AtomicInteger(0)
+      val logErrorCounter = new AtomicInteger(0)
+      val incLogHandled: (String, Option[Throwable]) => F[Unit] = (_, _) =>
+        m.eval {
+          val _ = logHandledCounter.incrementAndGet()
+        }
+      val incLogException: (String, Throwable) => F[Unit] = (_, _) =>
+        m.eval {
+          val _ = logErrorCounter.incrementAndGet()
+        }
+      val silentEndpoint1 = endpoint.get
+        .in("silent_hello_2827")
+        .out(stringBody)
+
+      val silentEndpoint2 = endpoint.get
+        .in("silent_hello2_2827")
+        .out(stringBody)
+
+      val serverLog = DefaultServerLog(
+        doLogWhenReceived = _ => m.unit(()),
+        doLogAllDecodeFailures = (_, _) => m.unit(()),
+        noLog = m.unit(()),
+        doLogWhenHandled = incLogHandled,
+        doLogExceptions = incLogException
+      )
+        .ignoreEndpoints(Seq(silentEndpoint1, silentEndpoint2))
+      testServer(
+        "Log events with provided functions, skipping certain endpoints",
+        NonEmptyList.of(
+          route(
+            List(
+              endpoint.get
+                .in("hello_2827")
+                .out(stringBody)
+                .serverLogic[F](_ => pureResult("result-hello".asRight[Unit])),
+              silentEndpoint1
+                .serverLogic[F](_ => pureResult("result-silent_hello".asRight[Unit])),
+              silentEndpoint2
+                .serverLogic[F](_ => m.error(new Exception("Boom!")))
+            ),
+            (_: CustomiseInterceptors[F, OPTIONS])
+              .serverLog(serverLog)
+          )
+        )
+      ) { (backend, baseUri) =>
+        basicStringRequest.get(uri"$baseUri/hello_2827").send(backend).map { _ =>
+          logHandledCounter.get() shouldBe 1
+        } >>
+          basicStringRequest.get(uri"$baseUri/silent_hello_2827").send(backend).map { _ =>
+            logHandledCounter.get() shouldBe 1
+            logErrorCounter.get() shouldBe 0
+          } >>
+          basicStringRequest.get(uri"$baseUri/silent_hello2_2827").send(backend).map { _ =>
+            logHandledCounter.get() shouldBe 1
+            logErrorCounter.get() shouldBe 1
+          } >>
+          basicStringRequest.get(uri"$baseUri/hello_2827").send(backend).map { _ =>
+            logHandledCounter.get() shouldBe 2
+            logErrorCounter.get() shouldBe 1
+          }
       }
     }
   )
